+++ CHANGELOG.md
@@ -16,6 +16,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `IsValidDNA` strategies for non-ambiguous bases
 - Exposed `AlignmentStates`, a struct for incrementally building up CIGAR strings
 - Add `invert` method to `Alignment`
+- Added `sw_score_from_path` to obtain the local smith waterman alignment score from a CIGAR string
 
 ### Changed